パーミッション挙動変更 #36 #26
