+++ app/src/main/java/com/example/walkly/MapsActivity.kt
@@ -27,4 +27,5 @@ class MapsActivity : AppCompatActivity() {
             .findFragmentById(R.id.map) as SupportMapFragment
         mapFragment.getMapAsync(mapCallback)
     }
+
 }
